Handle PyObject_Unicode better and reinstate key requirement being ASCII
PyObject_Unicode in Python27 handled converting many input types to unicode
previous version of this change just checked for unicode and proceeded from there
This approach satisfied testing suite, however this is a change to input spec
in cases not covered by current tests. In order to address this the new approach
uses repr of object if not unicode, this is the same functionality
as PyObject_Unicode used.

Ascii keys are also reinstated as this could be an external requirement.