Fix tests that rely on a specific System Default Language
FormatNumber does not take a languageID as a parameter, but rather always formats depending on the System Default Language. This caused tests to fail when running on a machine that had a System Default Language that used different separators for numbers. Since this formatting behavior has been present since time immemorial it has become the unofficial standard for the behavior of the system, so instead of changing the system to match the tests, I'm updating the tests to match the observed behavior of the system.