Fix test properties on Windows
A semicolon was escaped where it shouldn't have been, resulting in the BUILDFLAVOR part of the string being added to the PATH. Removing the escape character so that BUILDFLAVOR gets set as its own variable. Since we only have one thing on our path we can use the same value on macOS and Windows.