Use `size_t` for wrap point list count
This variable was declared as int, but then it gets cast it to int and a check exists which checks if its value is greater than INT_MAX. This makes no sense. The Size() method of WrapPointList implicitly casts this value to size_t. Therefore, changing the type to size_t and not casting this variable to another type once it has been set.