Add missing `nullptr` checks when creating a `WordWrapList`
Currently, on TQ there is a corp description that contains "invalid" Unicode characters. These characters cause the underlying sys call to fail. There was no guard for this on macOS, so logic would proceed with a nullptr that then resulted in a crash. The test case resembles what caused the crash.

Fixes https://ccpgames.atlassian.net/browse/PLAT-9947